updated jenkinsfile

+++ Jenkinsfile
@@ -1,4 +1,3 @@
-
 pipeline {
     agent {
     node {
@@ -12,13 +11,17 @@ pipeline {
         stage('build') {
             steps {
                 echo 'building...'
+
                 script{
-                    withCredentials([usernamePassword(credentialsId: 'docker-hub-cred', usernameVariable: 'USERNAME', passwordVariable: 'PASSWORD')]) {
-                    sh '''    
-                        docker login -u ${USERNAME} -p ${PASSWORD}
-                        docker build . -t andrewanter/bakehouse:${BUILD_NUMBER}
-                        docker push andrewanter/bakehouse:v${BUILD_NUMBER}
-                    '''
+                    if( ${BRANCH_NAME} == "master" ){
+                        withCredentials([usernamePassword(credentialsId: 'docker-hub-cred', usernameVariable: 'USERNAME', passwordVariable: 'PASSWORD')]) {
+                        sh '''    
+                            docker login -u ${USERNAME} -p ${PASSWORD}
+                            docker build . -t andrewanter/bakehouse:${BUILD_NUMBER}
+                            docker push andrewanter/bakehouse:v${BUILD_NUMBER}
+                            echo ${BUILD_NUMBER} > ../build.txt
+                        '''
+                        }
                     }
                 }
                 
@@ -27,15 +30,20 @@ pipeline {
         stage('deploy') {
             steps {
                 echo 'deploying to localhost'
+
                 script{
-                    withCredentials([file(credentialsId: 'kube-config', variable: 'KUBECONFIG')]) {
-                    sh '''
-                        mv Deployment/deploy.yaml Deployment/deploy.yaml.tmp
-                        cat Deployment/deploy.yaml.tmp | envsubst > Deployment/deploy.yaml
-                        rm -f Deployment/deploy.yaml.tmp
-                        kubectl apply -f Deployment --kubeconfig ${KUBECONFIG}
-                        
-                    '''
+
+                    if( ${BRANCH_NAME} == "master" ){
+                        withCredentials([file(credentialsId: 'kube-config', variable: 'KUBECONFIG')]) {
+                        sh '''
+                            export BUILD_NUMBER=${cat ../build.txt}
+                            mv Deployment/deploy.yaml Deployment/deploy.yaml.tmp
+                            cat Deployment/deploy.yaml.tmp | envsubst > Deployment/deploy.yaml
+                            rm -f Deployment/deploy.yaml.tmp
+                            kubectl apply -f Deployment --kubeconfig ${KUBECONFIG}
+                            
+                        '''
+                        }
                     }
                 }   
             }
